Move exceptions to root namespace
