+++ CliWrap.Tests/ValidationSpecs.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using CliWrap.Buffered;
-using CliWrap.Exceptions;
 using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
