+++ CliWrap/Buffered/BufferedCommandExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Text;
 using System.Threading;
-using CliWrap.Exceptions;
 
 namespace CliWrap.Buffered;
